// src/app/Model/Request/ShipmentData.ts
export interface ShipmentData {
    NoShip?: string;
    PO?: string
    Customer?: string 
    MohId?: string
    ProdCode?: string
    Lot?: string	
    QtyTotal?: number
    qtyMeerpTotal?: number
    Description?: string // Traveler
    ShipNica?: string
    SentNica?: number
    Shipped?: number
    PermitShip?: number
    Operator?: string
    lbapp?: string
    lbgen?: string
}